filter empty entries
When using this attribute in a composer form concrete5 tries to autosave periodically. If there is not a selection for this attribute, the save will fail with a 500 error due to trying to save a null entry. By filtering the array before saving, null and empty selections are not passed along to the createAttributeValue() method